algo/src/test/java/com/hartwig/actin/algo/evaluation/molecular/GeneHasActivatingMutationTest.java
package com.hartwig.actin.algo.evaluation.molecular;

import static org.junit.Assert.assertEquals;

import com.hartwig.actin.TestDataFactory;
import com.hartwig.actin.algo.datamodel.Evaluation;

import org.junit.Test;

public class GeneHasActivatingMutationTest {

    @Test
    public void canEvaluate() {
        GeneHasActivatingMutation function = new GeneHasActivatingMutation("gene 1");

        assertEquals(Evaluation.FAIL, function.evaluate(TestDataFactory.createMinimalTestPatientRecord()));

        assertEquals(Evaluation.FAIL, function.evaluate(MolecularTestFactory.withActivatedGene("gene 2")));
        assertEquals(Evaluation.PASS, function.evaluate(MolecularTestFactory.withActivatedGene("gene 1")));
    }
}